Slow down with requests if we're approaching the rate limit
GitHub apparently chose to bother client with some nonstandard headers
for indicating that client should should down instead of employing
dedicated network hardware for the job.

Even better, they abuse 403 response instead of 503 with Retry-After
header which would make it possible for the HTTP client library take care
of the delays.

Instead of waiting after each request, we chose to do the first half of
minute's requests without delay and wait twice as much after the rest,
so that we wouldn't add an unnecessary delay for short sessions. The
algorithm could probably be improved to dynamically increase the delay.